Revamped animation system and added idle animations
Now the animations store rects related to the whole link spritesheet. When changing between frames now the animation only change the texture rect of the sprite given by the FrameRect stored, saving a lot of memory. In this version Link can look up, left, down and right and if the user doesn't introduce any input Link will play the idle animation for the direction he's facing.