Added a ton of things
The zoom now works quite fine, duplicating the zoom factor every time you scroll. Made grid to start with matrix-like collision system. Now it's done by software but I tried to do it in the GPU (that's why the shaders are there) unsuccessfully. Damn you, SFML. The grid can be toggled with the button (in this version). TODO: customize the grid parameters and start the collision system