Added InputText class
The InputText adjusts the characters size depending on the height of the box. It accepts characters even in caps and you can move between characters with the direction arrows. Still need to work on the right side overflow of the text. Also changed the configuration .lua file to add chrono to the tool