Big update. Added movement to Link
Now Link moves freely through the level. The raw input processing (catching the input straight from the window) has been moved to Link's processInput(), because for the movement it's needed to keep track of all the moving directions the player is pressing. These directions are stored in a boolean vector that is used to determine the movement direction